Update range check orientation handling again.
This resolves #11, hopefully for the last time. The orientation is not a property of the chart
but instead is a property of the environment in which it is printed. Thus, it doesn't make much
sense to record orientation as a property of the chart. I disabled "sseq range check sideways"
and "sseq range check standrd" as options to sseqdata environments and as default settings.
Whether the chart should be range checked at all can make sense as a chart setting or most
reasonably as a global setting.

I added new commands \SseqOrientationNormal, \SseqOrientationSideways, and \SseqOrientationToggle
to indicate whether the current environment is normal or sideways. These modify a normal latex
boolean which is not registered with sseqloadstore at all, so we won't see the issues we had before
with chart-associated values overriding the default set by the sideways environment.